Long ids

+++ src/main/java/guldilin/dto/ActiveSubstanceDTO.java
@@ -1,12 +1,11 @@
 package guldilin.dto;
 
-
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 public class ActiveSubstanceDTO {
 
-    private Integer id;
+    private Long id;
 
     @NotNull
     @NotBlank
@@ -14,11 +13,11 @@ public class ActiveSubstanceDTO {
 
     private String description;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
