+++ src/main/java/guldilin/dto/MedicamentClassDTO.java
@@ -1,13 +1,11 @@
 package guldilin.dto;
 
-import guldilin.model.MedicamentGroup;
-
 import javax.validation.constraints.NotNull;
 
 public class MedicamentClassDTO {
 
 
-    private Integer id;
+    private Long id;
 
     private Integer parentGroup;
 
@@ -16,11 +14,11 @@ public class MedicamentClassDTO {
 
     private String description;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
