+++ src/main/java/guldilin/dto/WorkerDTO.java
@@ -1,12 +1,10 @@
 package guldilin.dto;
 
-import guldilin.model.WorkerRole;
-
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 
 public class WorkerDTO implements Serializable {
-    private Integer id;
+    private Long id;
 
     @NotNull
     private String name;
@@ -18,11 +16,11 @@ public class WorkerDTO implements Serializable {
 
     private String email;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
